+++ tests/test_checker/test_noqa.py
@@ -200,6 +200,7 @@ SHOULD_BE_RAISED = types.MappingProxyType({
     'WPS522': 1,
     'WPS523': 1,
     'WPS524': 1,
+    'WPS525': 2,
 
     'WPS600': 1,
     'WPS601': 1,